% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessSVG.R
\name{preprocessSVG}
\alias{preprocessSVG}
\title{preprocessSVG}
\usage{
preprocessSVG(spe, in_tissue = TRUE, filter_genes = 20, filter_mito = TRUE)
}
\arguments{
\item{spe}{\code{SpatialExperiment} Input data, assumed to be a 
\code{\link{SpatialExperiment}} object containing an assay named 
\code{counts} containing raw unique molecular identifier (UMI) counts, and 
spatial coordinates stored in the \code{spatialCoords} slot.}

\item{in_tissue}{\code{logical} Whether to keep only spots over tissue, 
identified by the column \code{in_tissue} in the \code{spatialData} slot. 
Default = TRUE.}

\item{filter_genes}{\code{integer} Whether to filter low-expressed genes 
according to total unique molecular identifier (UMI) counts per gene 
across spatial coordinates. If an integer value is provided, genes with 
less than or equal to this number of total UMI counts across spatial 
coordinates will be removed. Assumes \code{spe} contains an assay named 
\code{counts} containing raw UMI counts. Default = 20. Set to NULL to 
disable.}

\item{filter_mito}{\code{logical} Whether to filter mitochondrial genes. 
Assumes the \code{rowData} slot of \code{spe} contains a column named 
\code{gene_name}, which can be used to identify mitochondrial genes. 
Default = TRUE. Set to FALSE to disable.}
}
\value{
Returns a \code{SpatialExperiment} object that can be provided to 
  \code{\link{nnSVG}}.
}
\description{
Preprocessing steps to prepare input data for nnSVG
}
\details{
Convenience function to run several preprocessing steps to prepare input 
data object for nnSVG. This function is designed for data from the 10x 
Genomics Visium platform, and is used for examples in the nnSVG package.

In general, the code in this function will need to be adapted for a given 
dataset, so we recommend running the steps individually instead of using 
this function. The steps are described in detail in our online book 
\code{Orchestrating Spatially Resolved Transcriptomics Analysis with 
Bioconductor (OSTA)}.
}
\examples{
library(SpatialExperiment)
library(STexampleData)

spe <- Visium_humanDLPFC()

# set seed for reproducibility
set.seed(123)
spe <- preprocessSVG(spe)

spe

}
